Added partition merging functionality to optimization command
